First attempt at moving SalesForce API code to DAO.
