+++ orcid-core/src/main/java/org/orcid/core/manager/impl/SalesForceMembersListCacheEntryFactory.java
@@ -18,7 +18,7 @@ package org.orcid.core.manager.impl;
 
 import javax.annotation.Resource;
 
-import org.orcid.core.manager.SalesForceManager;
+import org.orcid.core.salesforce.dao.SalesForceDao;
 
 import net.sf.ehcache.constructs.blocking.CacheEntryFactory;
 
@@ -30,11 +30,11 @@ import net.sf.ehcache.constructs.blocking.CacheEntryFactory;
 public class SalesForceMembersListCacheEntryFactory implements CacheEntryFactory {
 
     @Resource
-    private SalesForceManager salesForceManager;
+    private SalesForceDao salesForceDao;
 
     @Override
     public Object createEntry(Object key) throws Exception {
-        return salesForceManager.retrieveFreshMembers();
+        return salesForceDao.retrieveFreshMembers();
     }
 
 }
